Update jquery.convform.js
This a new commit to add a "repeat" functionnality. If the user wants to repeat a part of the form, now it's possible. 

For that 3 new attribute there added. 
        - conv-initiate-repeat="true";
        - conv-start-repeat="true";
        - conv-stop-repeat="true"

The attribute "conv-initiate-repeat" can be set at any time of the form. This is from where the form will restart.

To repeat the section of the form, it's necessary to use "conv-start-repeat" and "conv-stop-repeat" parameter in a select. 
If the user click on the option where:
           - "conv-start-repeat" is set, the form will repeat from the question where was set "conv-initiate-repeat" parameter. 
           - "conv-stop-repeat" is set, the form will continue from where it should. After the repeated section.

Example and limitation : 

The element must have a name using an "array format" like : 
- name="item[idx][productName]" . 

in this example idx will be incremented automatically by the plugin. You can put the value that you want but it's pointless as it will be replaced by an int (0, 1, 2, 3, 4 ,5 ,6 ...).


<input name="item[idx][name]" type="text" conv-question="what is the name of the item" conv-initiate-repeat="true">
<select name="item[idx][kind]" conv-question="is it a car or a bike ?"  required="required">
    <option value="car">car</option>
    <option value="bike">bike</option>
</select>
<input name="item[idx][price]" type="text" conv-question="what is the price ?">
<select name="newItem" conv-question="Do you want to add a new item ?"  required="required">
    <option value="0" conv-start-repeat="true" >yes</option>
    <option value="1" conv-stop-repeat="true" >no</option>
</select>
<input  type="text" conv-question="what is the name of the item"  no-answer="true">


on server side you will receive an array like that : 
item[0][name]=".....",
item[0][kind]=".....",
item[0][price]=".....",
item[1][name]=".....",
item[2][kind]=".....",
item[3][price]=".....",

Br, 
Fas3r.